requirements.txt
browser-use>=0.1.0
langchain-openai>=0.2.0
playwright>=1.49.0
python-dotenv>=1.0.0

